Minor code improvements
- programmatically set toolbar color instead of in xml
- functions used in multiple activities and fragment moved to BaseActivity, implemented an interface to call those methods from fragment
