+++ app/src/main/res/values/strings.xml
@@ -16,5 +16,7 @@
     <!--Share Preference-->
     <string name="shared_preference_file">com.pixellore.checklist.PREFERENCE_FILE</string>
     <string name="shared_pref_app_theme_key">KEY_THEME</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
