+++ app/src/main/res/values/styles.xml
@@ -27,16 +27,4 @@
     <style name="actionOverflowButtonStyle" parent="@style/Widget.AppCompat.ActionButton.Overflow">
         <item name="android:tint">?attr/colorOnPrimary</item>
     </style>
-
-    <!--Toolbar Style-->
-    <style name="ToolbarStyle" parent="ThemeOverlay.AppCompat.ActionBar">
-        <item name="android:background">?attr/colorPrimary</item>
-        <item name="titleTextColor">?attr/colorOnPrimary</item>
-        <item name="colorControlNormal">?attr/colorOnPrimary</item>
-    </style>
-
-    <style name="ToolbarPopUpMenu" parent="ThemeOverlay.AppCompat.Light">
-        <item name="android:textColor">?attr/colorOnPrimary</item>
-        <item name="android:itemBackground">?attr/colorPrimary</item>
-    </style>
 </resources>
